feature/GAL-345# 100% helper func coverage
