+++ package.json
@@ -25,6 +25,7 @@
     "csv": "^6.3.6",
     "csv-parser": "^3.0.0",
     "csv-stringify": "^6.4.4",
-    "loglevel": "^1.8.1"
+    "loglevel": "^1.8.1",
+    "nhs-mesh-client": "^1.0.8"
   }
 }
